pyproject.toml
[tool.poetry]
name = "blogapi"
version = "0.1.0"
description = ""
authors = ["Aydar Gimadiev <gima.aydar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1"
djangorestframework = "3.11.0"
dj-rest-auth = "1.1.0"
django-allauth = "~=0.42.0"
PyYAML = "5.3.1"
uritemplate = "3.0.1"
drf-yasg = "1.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
